requirements.txt
# requirements.txt

annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.7
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
shellingham==1.5.4
sniffio==1.3.1
typing_extensions==4.12.2
ujson==5.10.0
websockets==13.1
